+++ src/Serializer/AbstractItemNormalizer.php
@@ -14,7 +14,9 @@ declare(strict_types=1);
 namespace ApiPlatform\Core\Serializer;
 
 use ApiPlatform\Core\Api\IriConverterInterface;
+use ApiPlatform\Core\Api\IriPlainIdentifierAwareConverterInterface;
 use ApiPlatform\Core\Api\OperationType;
+use ApiPlatform\Core\Api\PlainIdentifierConverterInterface;
 use ApiPlatform\Core\Api\ResourceClassResolverInterface;
 use ApiPlatform\Core\DataProvider\ItemDataProviderInterface;
 use ApiPlatform\Core\Exception\InvalidArgumentException;
@@ -43,6 +45,9 @@ abstract class AbstractItemNormalizer extends AbstractObjectNormalizer
 
     protected $propertyNameCollectionFactory;
     protected $propertyMetadataFactory;
+    /**
+     * @var IriPlainIdentifierAwareConverterInterface|IriConverterInterface $iriConverter
+     */
     protected $iriConverter;
     protected $resourceClassResolver;
     protected $propertyAccessor;
@@ -59,6 +64,9 @@ abstract class AbstractItemNormalizer extends AbstractObjectNormalizer
         $this->iriConverter = $iriConverter;
         $this->resourceClassResolver = $resourceClassResolver;
         $this->propertyAccessor = $propertyAccessor ?: PropertyAccess::createPropertyAccessor();
+        if (null !== $itemDataProvider) {
+            @trigger_error(sprintf('Passing a %s is deprecated since 2.3 and will not be possible in 3.0.', ItemDataProviderInterface::class));
+        }
         $this->itemDataProvider = $itemDataProvider;
         $this->allowPlainIdentifiers = $allowPlainIdentifiers;
 
@@ -309,9 +317,9 @@ abstract class AbstractItemNormalizer extends AbstractObjectNormalizer
 
         if (!\is_array($value)) {
             // repeat the code so that IRIs keep working with the json format
-            if (true === $this->allowPlainIdentifiers && $this->itemDataProvider) {
+            if (true === $this->allowPlainIdentifiers && $this->iriConverter instanceof IriPlainIdentifierAwareConverterInterface) {
                 try {
-                    return $this->itemDataProvider->getItem($className, $value, null, $context + ['fetch_data' => true]);
+                    return $this->iriConverter->getItemFromIri($this->iriConverter->getIriFromPlainIdentifier($value, $className), $context + ['fetch_data' => true]);
                 } catch (ItemNotFoundException $e) {
                     throw new InvalidArgumentException($e->getMessage(), $e->getCode(), $e);
                 } catch (InvalidArgumentException $e) {